Add additional app launch scenarios for OOM tracking
Summary:
OOM tracking is based off of the blog post from Facebook https://code.fb.com/ios/reducing-fooms-in-the-facebook-ios-app/

This requires a process-of-elimination to determine if a launch is due to an OOM crash by marking every legitimate reason that an application could start. The article did not cover ways that the app may be launched from the OS which are background fetches and silent push notifications currently.

  New additions in this PR
  - Track app launches for silent notifications and background fetch
  - Don't report OOMs if the previous session was for these reasons.